Compress some fields in mempage.
Shrink mempage a bit to conserve memory.

We only need 32bit of counter.
Move other fields around to avoid padding.

+++ memdb.h
@@ -3,7 +3,7 @@
 
 struct err_type {
 	struct leaky_bucket bucket;
-	unsigned long count;
+	unsigned count;
 };
 
 enum printflags {
